feat: Upgrade app to Laravel 13 and PHP 8.4 (#251)

+++ database/migrations/2025_06_04_100009_create_plans_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('code')->nullable()->unique();
             $table->string('description')->nullable();
             $table->integer('days')->nullable();
-            $table->float('amount')->nullable();
+            $table->decimal('amount', 12, 2)->nullable();
             $table->enum('status', ['active', 'inactive'])->default('active')->nullable();
             $table->softDeletes();
             $table->timestamps();
